rag/requirements.txt
flask==3.0.3
requests==2.32.3
scikit-learn==1.5.1
numpy==1.26.4
python-dotenv==1.0.1
